Deduplicate problem half done
debug exp:
if base class is a template, and derived class is an specified template,
it seems the derived class can not override virtual functions in base
class. If the derived class is switched to template mode, then
everything's OK.
My program works well with windows code::blocks, but crash immediately
with linux(OJ or server) by run time error(access invalid memory). After
painful searching on the Internet(which is remarkably useless) and
debugging both under windows and linux, I found the reason is because
different mechanisms of rand(). The mothod I used is to cout at
important position(again).
it tells me why I shoud be cautious with rand() when implementing Hash
and the handy method of debugging